# src/openapi.yaml
openapi: 3.0.2
info:
  x-logo:
    href: https://kortxyz.github.io/favicon.ico
  title: OGCAPI
  version: 0.0.1
  description: Implemations of OGC API
  contact:
    name: Tino Kastbjerg Stigsen
    email: info@kort.xyz
    url: https://kort.xyz
  license:
    name: CC-BY 4.0 license
    url: https://creativecommons.org/licenses/by/4.0/
servers:
  - url: https://localhost:3000
...
paths:
...
  /collections/{collectionId}:
    get:
      operationId: getCollection
      tags:
        - Capabilities
      summary: get collection metadata
      description: Show details about a collection available in the dataset.
      parameters:
        - $ref: "#/components/parameters/fCommon"
        - $ref: "#/components/parameters/collectionId"
      responses:
        200:
          description: collections metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/collection"
        400:
          $ref: "#/components/responses/ServerError"
      security:
        - {}
...
  /collections/{collectionId}/map/tiles:
    get:
      operationId: getCollectionMapTilesets
      tags:
        - Collection Tiles
      summary: get Map Tilesets
      description: Fetch a list of tilematrixsets
      parameters:
        - $ref: "#/components/parameters/collectionId"
      responses:
        200:
          description: List of available tilesets.
          content:
            text/html:
              schema:
                type: string
            application/json:
              schema:
                type: object
                required:
                  - tilesets
                properties:
                  links:
                    type: array
                    items:
                      $ref: "#/components/schemas/link"
                  tilesets:
                    type: array
                    items:
                      $ref: "#/components/schemas/tileSet"
        400:
          $ref: "#/components/responses/ServerError"
      security:
        - {}
  /collections/{collectionId}/map/tiles/{tileMatrixSetId}:
    get:
      operationId: getCollectionMapTileset
      tags:
        - Collection Tiles
      summary: get Tileset
      description: Fetch description of a specific tilematrixset
      parameters:
        - $ref: "#/components/parameters/collectionId"
        - $ref: "#/components/parameters/tileMatrixSetId"
        - $ref: "#/components/parameters/fTileset"
      responses:
        200:
          description: Description of the tileset
          content:
            text/html:
              schema:
                type: string
            application/tilejson:
              schema:
                type: object
            application/json:
              schema:
                $ref: "#/components/schemas/tileSet"
              examples:
                tileSet:
                  $ref: "#/components/examples/tileSet"

        400:
          $ref: "#/components/responses/ServerError"

      security:
        - {}
  /collections/{collectionId}/map/tiles/{tileMatrixSetId}/{tileMatrix}/{tileCol}/{tileRow}: 
    get:
      operationId: getCollectionMapTile
      tags:
        - Collection Tiles
      summary: get MapTile
      description: Fetches a single maptile from a collection
      parameters:
        - $ref: "#/components/parameters/collectionId"
        - $ref: "#/components/parameters/tileMatrixSetId"
        - $ref: "#/components/parameters/tileMatrix"
        - $ref: "#/components/parameters/tileRow"
        - $ref: "#/components/parameters/tileCol"
      responses:
        200:
          description: Tile
          content:
            image/*:
              schema:
                type: string
                format: binary
        400:
          $ref: "#/components/responses/ServerError"
      security:
        - {}
...
  /tileMatrixSets:
    get:
      tags:
        - Tile Matrix Sets
      summary: fetch all available tile matrix sets (tiling schemes)
      operationId: getTileMatrixSets
      parameters:
        - $ref: "#/components/parameters/fCommon"
      responses:
        200:
          description: An array of tile matrix sets (tiling schemes).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tileMatrixSets"
        400:
          $ref: "#/components/responses/ServerError"
      security:
        - {}
  /tileMatrixSets/{tileMatrixSetId}:
    get:
      tags:
        - Tile Matrix Sets
      summary: fetch a tile matrix sets (tiling scheme) by id
      operationId: getTileMatrixSet
      parameters:
        - $ref: "#/components/parameters/fCommon"
        - $ref: "#/components/parameters/tileMatrixSetId"
      responses:
        200:
          description: tile matrix sets (a tiling scheme).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tileMatrixSet"
        400:
          $ref: "#/components/responses/ServerError"
      security:
        - {}
  /collections/{collectionId}/tiles:
    get:
      operationId: getCollectionTilesets
      tags:
        - Collection Tiles
      summary: get Tilesets
      description: Fetch a list of tilematrixsets
      parameters:
        - $ref: "#/components/parameters/collectionId"
      responses:
        200:
          description: List of available tilesets.
          content:
            text/html:
              schema:
                type: string
            application/json:
              schema:
                type: object
                required:
                  - tilesets
                properties:
                  links:
                    type: array
                    items:
                      $ref: "#/components/schemas/link"
                  tilesets:
                    type: array
                    items:
                      $ref: "#/components/schemas/tileSet"
        400:
          $ref: "#/components/responses/ServerError"
      security:
        - {}
  /collections/{collectionId}/tiles/{tileMatrixSetId}:
    get:
      operationId: getCollectionTileset
      tags:
        - Collection Tiles
      summary: get Tileset
      description: Fetch description of a specific tilematrixset
      parameters:
        - $ref: "#/components/parameters/collectionId"
        - $ref: "#/components/parameters/tileMatrixSetId"
        - $ref: "#/components/parameters/fTileset"
      responses:
        200:
          description: Description of the tileset
          content:
            text/html:
              schema:
                type: string
            application/tilejson:
              schema:
                type: object
            application/json:
              schema:
                $ref: "#/components/schemas/tileSet"
              examples:
                tileSet:
                  $ref: "#/components/examples/tileSet"

        400:
          $ref: "#/components/responses/ServerError"

      security:
        - {}
  /collections/{collectionId}/tiles/{tileMatrixSetId}/{tileMatrix}/{tileRow}/{tileCol}:
    get:
      operationId: getCollectionTile
      tags:
        - Collection Tiles
      summary: get Tile
      description: Fetches a single tile from a collection
      parameters:
        - $ref: "#/components/parameters/collectionId"
        - $ref: "#/components/parameters/tileMatrixSetId"
        - $ref: "#/components/parameters/tileMatrix"
        - $ref: "#/components/parameters/tileRow"
        - $ref: "#/components/parameters/tileCol"
      responses:
        200:
          description: Tile
          content:
            image/*:
              schema:
                type: string
                format: binary
        400:
          $ref: "#/components/responses/ServerError"
      security:
        - {}
  /tiles:
    get:
      operationId: getTilesets
      tags:
        - Dataset Tiles
      summary: get Tilesets
      description: Fetch a list of tilematrixsets
      responses:
        200:
          description: List of available tilesets.
          content:
            text/html:
              schema:
                type: string
            application/json:
              schema:
                type: object
                required:
                  - tilesets
                properties:
                  links:
                    type: array
                    items:
                      $ref: "#/components/schemas/link"
                  tilesets:
                    type: array
                    items:
                      $ref: "#/components/schemas/tileSet"
        400:
          $ref: "#/components/responses/ServerError"
      security:
        - {}
  /tiles/{tileMatrixSetId}:
    get:
      operationId: getTileset
      tags:
        - Dataset Tiles
      summary: get Tileset
      description: Fetch description of a specific tilematrixset
      parameters:
        - $ref: "#/components/parameters/tileMatrixSetId"
        - $ref: "#/components/parameters/fTileset"
      responses:
        200:
          description: Description of the tileset
          content:
            text/html:
              schema:
                type: string
            application/tilejson:
              schema:
                type: object
            application/json:
              schema:
                $ref: "#/components/schemas/tileSet"
              examples:
                tileSet:
                  $ref: "#/components/examples/tileSet"

        400:
          $ref: "#/components/responses/ServerError"

      security:
        - {}
  /tiles/{tileMatrixSetId}/{tileMatrix}/{tileRow}/{tileCol}:
    get:
      operationId: getTile
      tags:
        - Dataset Tiles
      summary: get Tile
      description: Fetches a single tile from a collection
      parameters:
        - $ref: "#/components/parameters/tileMatrixSetId"
        - $ref: "#/components/parameters/tileMatrix"
        - $ref: "#/components/parameters/tileRow"
        - $ref: "#/components/parameters/tileCol"
      responses:
        200:
          description: Tile
          content:
            image/*:
              schema:
                type: string
                format: binary
        400:
          $ref: "#/components/responses/ServerError"
      security:
        - {}
...
components:
  parameters:
    fCommon:
      name: f
      in: query
      description: >-
        Select the output format of the response. If no value is provided, the
        standard HTTP rules apply, i.e., the accept header will be used to
        determine the format.
      required: false
      schema:
        type: string
        enum:
          - json
          - html
    fTileset:
      name: f
      in: query
      description: >-
        Select the output format of the response. If no value is provided, the
        standard HTTP rules apply, i.e., the accept header will be used to
        determine the format.
      required: false
      schema:
        type: string
        enum:
          - json
          - html
          - tilejson
...
    collectionId:
      name: collectionId
      in: path
      description: local identifier of a collection
      required: true
      schema:
        type: string
...
    tileMatrixSetId:
      name: tileMatrixSetId
      in: path
      description: ID of the tileMatrixset
      example: WebMercatorQuad
      required: true
      schema:
        type: string
    tileMatrix:
      name: tileMatrix
      in: path
      description: z value for the tile
      required: true
      schema:
        type: number
    tileRow:
      name: tileRow
      in: path
      description: x value for the tile
      required: true
      schema:
        type: number
    tileCol:
      name: tileCol
      in: path
      description: y value for the tile
      required: true
      schema:
        type: number
...
  schemas:
    link:
      required:
        - href
        - rel
        - type
      type: object
      properties:
        rel:
          type: string
        type:
          type: string
        title:
          type: string
        href:
          type: string
        templated:
          type: boolean
...
    collection:
      type: object
      required:
        - id
        - links
        - extent
      properties:
        id:
          description: identifier of the collection used, for example, in URIs
          type: string
        title:
          description: human readable title of the collection
          type: string
        description:
          description: a description of the features in the collection
          type: string
        links:
          type: array
          items:
            $ref: "#/components/schemas/link"
        extent:
          description: >-
            The extent of the features in the collection. In the Core only
            spatial and temporal extents are specified. Extensions may add
            additional members to represent other extents, for example, thermal
            or pressure ranges.
          type: object
          properties:
            spatial:
              description: The spatial extent of the features in the collection.
              type: object
              properties:
                bbox:
                  description: >-
                    One or more bounding boxes that describe the spatial extent
                    of the dataset. In the Core only a single bounding box is
                    supported. Extensions may support additional areas. If
                    multiple areas are provided, the union of the bounding boxes
                    describes the spatial extent.
                  type: array
                  minItems: 1
                  items:
                    description: >-
                      Each bounding box is provided as four or six numbers,
                      depending on whether the coordinate reference system
                      includes a vertical axis (height or depth):

                      * Lower left corner, coordinate axis 1 * Lower left
                      corner, coordinate axis 2 * Minimum value, coordinate axis
                      3 (optional) * Upper right corner, coordinate axis 1 *
                      Upper right corner, coordinate axis 2 * Maximum value,
                      coordinate axis 3 (optional)

                      The coordinate reference system of the values is WGS 84
                      longitude/latitude
                      (http://www.opengis.net/def/crs/OGC/1.3/CRS84) unless a
                      different coordinate reference system is specified in
                      `crs`.

                      For WGS 84 longitude/latitude the values are in most cases
                      the sequence of minimum longitude, minimum latitude,
                      maximum longitude and maximum latitude. However, in cases
                      where the box spans the antimeridian the first value
                      (west-most box edge) is larger than the third value
                      (east-most box edge).

                      If the vertical axis is included, the third and the sixth
                      number are the bottom and the top of the 3-dimensional
                      bounding box.

                      If a feature has multiple spatial geometry properties, it
                      is the decision of the server whether only a single
                      spatial geometry property is used to determine the extent
                      or all relevant geometries.
                    type: array
                    minItems: 4
                    maxItems: 6
                    items:
                      type: number
                    example:
                      - -180
                      - -90
                      - 180
                      - 90
                crs:
                  description: >-
                    Coordinate reference system of the coordinates in the
                    spatial extent (property `bbox`). The default reference
                    system is WGS 84 longitude/latitude. In the Core this is the
                    only supported coordinate reference system. Extensions may
                    support additional coordinate reference systems and add
                    additional enum values.
                  type: string
            temporal:
              description: The temporal extent of the features in the collection.
              type: object
              properties:
                interval:
                  description: >-
                    One or more time intervals that describe the temporal extent
                    of the dataset. The value `null` is supported and indicates
                    an open time intervall. In the Core only a single time
                    interval is supported. Extensions may support multiple
                    intervals. If multiple intervals are provided, the union of
                    the intervals describes the temporal extent.
                  type: array
                  minItems: 1
                  items:
                    description: >-
                      Begin and end times of the time interval. The timestamps
                      are in the coordinate reference system specified in `trs`.
                      By default this is the Gregorian calendar.
                    type: array
                    minItems: 2
                    maxItems: 2
                    items:
                      type: string
                      format: date-time
                      nullable: true
                    example:
                      - "2011-11-11T12:22:11Z"
                      - null
                trs:
                  description: >-
                    Coordinate reference system of the coordinates in the
                    temporal extent (property `interval`). The default reference
                    system is the Gregorian calendar. In the Core this is the
                    only supported temporal reference system. Extensions may
                    support additional temporal reference systems and add
                    additional enum values.
                  type: string
                  enum:
                    - http://www.opengis.net/def/uom/ISO-8601/0/Gregorian
                  default: http://www.opengis.net/def/uom/ISO-8601/0/Gregorian
        itemType:
          description: >-
            indicator about the type of the items in the collection (the default
            value is 'feature').
          type: string
          default: feature
        crs:
          description: the list of coordinate reference systems supported by the service
          type: string
...
    tileMatrixSets:
      required:
        - tileMatrixSets
      type: object
      properties:
        links:
          type: array
          items:
            $ref: "#/components/schemas/link"
        tileMatrixSets:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              title:
                type: string
              tileMatrixSetURI:
                type: string
              links:
                type: array
                items:
                  $ref: "#/components/schemas/link"
            example:
              id: MyWebMercatorQuad
              title: My Google Maps Compatible for the World
              links:
                - href: https://data.example.org/tileMatrixSet/MyWebMercatorQuad
                  rel: item
                  type: application/json
    tileMatrixSetLimits:             
      type: object
      required:
      - tileMatrix
      - minTileRow
      - maxTileRow
      - minTileCol
      - maxTileCol
      properties:
        tileMatrix:
          type: string
          example: '5'
        minTileRow:
          type: number
          format: integer
          minimum: 0
          example: 0
        maxTileRow:
          type: number
          format: integer
          minimum: 0
          example: 1
        minTileCol:
          type: number
          format: integer
          minimum: 0
          example: 3
        maxTileCol:
          type: number
          format: integer
          minimum: 0
          example: 4
    exception:
      type: object
      description: >-
        Information about the exception: an error code plus an optional
        description.
      required:
        - statusCode
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
      example:
        statusCode: 404
        error: Not Found
        message: Route GET:/tileMatrixSets not found
...
    tileMatrixSet:
      required:
        - id
      type: object
      properties:
        id:
          type: string
          description: Tile matrix set identifier,
          example: WebMercatorQuad
        title:
          type: string
          description: Title of this tile matrix set, normally used for display to a human
          example: Google Maps Compatible for the World
        crs:
          type: string
          description: Reference to one coordinate reference system (CRS)
          format: uri
          example: http://www.opengis.net/def/crs/EPSG/0/3857
        tileMatrixSetURI:
          type: string
          description: Reference to a well-known scale set
          example: http://www.opengis.net/def/tilematrixset/OGC/1.0/WebMercatorQuad
        tileMatrices:
          type: array
          description: Describes a scale level and its tile matrix
          items:
            $ref: "#/components/schemas/tileMatrix"
        orderedAxes:
          type: array
          description: Order of the axes.
          items:
            type: string
        links:
          type: array
          items:
            $ref: "#/components/schemas/link"
    tileSet:
      title: Tile Set Metadata
      description:
        A resource describing a tileset based on the OGC TileSet Metadata Standard.
        At least one of the 'TileMatrixSet',  or a link with 'rel' http://www.opengis.net/def/rel/ogc/1.0/tiling-scheme
      type: object
      required: [dataType, crs, links]
      properties:
        title:
          description: A title for this tileset
          type: string
        description:
          description: Brief narrative description of this tile set
          type: string
        dataType:
          allOf:
            - description: Type of data represented in the tileset
            - $ref: "#/components/schemas/dataType"
        crs:
          allOf:
            - description: Coordinate Reference System (CRS)
            - $ref: "#/components/schemas/crs"
        tileMatrixSetURI:
          description:
            Reference to a Tile Matrix Set on an offical source for Tile Matrix
            Sets such as the OGC NA definition server (http://www.opengis.net/def/tms/).
            Required if the tile matrix set is registered on an open official source.
          type: string
          format: uri
        links:
          description:
            "Links to related resources. Possible link 'rel' values are: 'http://www.opengis.net/def/rel/ogc/1.0/dataset'
            for a URL pointing to the dataset, 'item' for a URL template to get a tile;
            'alternate' for a URL pointing to another representation of the TileSetMetadata
            (e.g a TileJSON file); 'http://www.opengis.net/def/rel/ogc/1.0/tiling-scheme'
            for a definition of the TileMatrixSet; 'http://www.opengis.net/def/rel/ogc/1.0/geodata'
            for pointing to a single collection (if the tileset represents a single collection)"
          type: array
          items:
            $ref: "#/components/schemas/link"
        tileMatrixSetLimits:
          description: Limits for the TileRow and TileCol values for each TileMatrix in
            the tileMatrixSet. If missing, there are no limits other that the ones imposed
            by the TileMatrixSet. If present the TileMatrices listed are limited and the
            rest not available at all
          type: array
          items:
            $ref: '#/components/schemas/tileMatrixSetLimits'
        # epoch:
        #   description: Epoch of the Coordinate Reference System (CRS)
        #   type: number
        layers:
          minItems: 1
          type: array
          items:
            $ref: "#/components/schemas/geospatialData"
        boundingBox:
          allOf:
            - description: Minimum bounding rectangle surrounding the tile matrix set, in the supported CRS
            - $ref: "#/components/schemas/2DBoundingBox"
        centerPoint:
          allOf:
            - description:
                Location of a tile that nicely represents the tileset. Implementations
                may use this center value to set the default location or to present a representative
                tile in a user interface
            - $ref: "#/components/schemas/tilePoint"
        # style:
        #   allOf:
        #     - description: Style involving all layers used to generate the tileset
        #     - $ref: 'style.yaml'
        attribution:
          description: Short reference to recognize the author or provider
          type: string
        license:
          description: License applicable to the tiles
          type: string
        accessConstraints:
          description:
            Restrictions on the availability of the Tile Set that the user needs
            to be aware of before using or redistributing the Tile Set
          type: string
          enum:
            - unclassified
            - restricted
            - confidential
            - secret
            - topSecret
        keywords:
          description: keywords about this tileset
          type: array
          items:
            type: string
        version:
          description:
            Version of the Tile Set. Changes if the data behind the tiles has
            been changed
          type: string
        created:
          allOf:
            - description: When the Tile Set was first produced
            - $ref: "#/components/schemas/timeStamp"
        updated:
          allOf:
            - description: Last Tile Set change/revision
            - $ref: "#/components/schemas/timeStamp"
        pointOfContact:
          description:
            Useful information to contact the authors or custodians for the Tile
            Set
          type: string
        mediaTypes:
          description: Media types available for the tiles
          type: array
          items:
            type: string
    dataType:
      anyOf:
        - type: string
          enum:
            - map
            - vector
            - coverage
    crs:
      title: CRS
      description: Reference to one coordinate reference system (CRS)
      type: string
...
    2DBoundingBox:
      description: Minimum bounding rectangle surrounding a 2D resource in the CRS indicated elsewhere
      type: object
      required:
        - lowerLeft
        - upperRight
      properties:
        lowerLeft:
          $ref: "#/components/schemas/2DPoint"
        upperRight:
          $ref: "#/components/schemas/2DPoint"
        crs:
          $ref: "#/components/schemas/crs"
        orderedAxes:
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: string
...
    tilePoint:
      type: object
      required:
        - coordinates
      properties:
        coordinates:
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: number
        crs:
          allOf:
            - description: Coordinate Reference System (CRS) of the coordinates
            - $ref: "#/components/schemas/crs"
        tileMatrix:
          description: TileMatrix identifier associated with the scaleDenominator
          type: string
        scaleDenominator:
          description: Scale denominator of the tile matrix selected
          type: number
        cellSize:
          description: Cell size of the tile matrix selected
          type: number
...
  responses:
...
    ServerError:
      description: A server error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/exception"
  examples:
    tileSet:
      value:
        title: Test
        dataType: map
        crs: "ESPG:4326"
        links:
          - href: http://data.example.com/id/building/123
            rel: canonical
            type: text/html
            title: canonical URI of the building
      summary: A sample tileset
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic