# src/routes/tiles.js
import * as model from "../database/tiles.js"
import * as helpers from "../helpers/tiles.js"
import * as templates from "../templates/tiles.js"

import * as modelCommon from "../database/common.js"

import { filetypemime } from 'magic-bytes.js';

async function getTilesets(req, reply) {
  const db = req.db || req.server.db;
  const baseurl = req.params.dataset ? [req.server.baseurl, req.params.dataset].join("/") : req.server.baseurl;

  const { contentType } = req;

  const collection = modelCommon.getCollection(db, "tiles");
  if(!collection) return reply.callNotFound();

  if (collection.data_type == "tiles") return reply.callNotFound();
  if (contentType == "html") return reply.view("tilesets", {baseurl });

  reply.send(templates.collectionTileSets(baseurl, collection));
};

async function getTileset(req, reply) {
  const db = req.db || req.server.db;
  const baseurl = req.params.dataset ? [req.server.baseurl, req.params.dataset].join("/") : req.server.baseurl;

  const { contentType } = req;

  const collection = modelCommon.getCollection(db, "tiles")

  if (contentType == "html") return reply.view("tileset", { baseurl });

  const layers = collection.data_type == 'vector-tiles' ? await model.getVectorTilesSpec(db, collection.name) : [{"id": collection.name,"dataType":'vector'}];
  
  if (contentType == "json") reply.send(templates.collectionTileSet(baseurl, collection, layers));

};

async function getTile(req, reply) {
  const { tileMatrix, tileRow, tileCol } = req.params;
  const db = req.db || req.server.db;

  let tile = await model.getCollectionTile(db,"tiles", tileMatrix, tileRow, tileCol);
  
  if (tile) {
    if (typeof tile == "object") tile = Buffer.from(tile)
    const gzip = await helpers.isGzip(tile)

    reply
      .header('Content-Type', 'application/vnd.mapbox-vector-tile')
      .header('Content-Encoding', gzip ? 'gzip' : 'none')
      .send(tile)
  }
  else {
    reply.status(404).send()
  }

};

async function getCollectionTilesets(req, reply) {
  const db = req.db || req.server.db;
  const baseurl = req.params.dataset ? [req.server.baseurl, req.params.dataset].join("/") : req.server.baseurl;

  const { contentType } = req;
  const { collectionId } = req.params;

  const collection = modelCommon.getCollection(db, collectionId);

  if (collection.data_type == "tiles") return reply.callNotFound();
  if (contentType == "html") return reply.view("collectiontilesets", {baseurl, collectionId });

  reply.send(templates.collectionTileSets(baseurl, collection));
};

async function getCollectionTileset(req, reply) {
  const db = req.db || req.server.db;
  const baseurl = req.params.dataset ? [req.server.baseurl, req.params.dataset].join("/") : req.server.baseurl;

  const { contentType } = req;
  const { collectionId } = req.params;

  const collection = modelCommon.getCollection(db, collectionId)

  if (contentType == "html") return reply.view("tileset", { baseurl, collectionId });

  const layers = collection.data_type == 'vector-tiles' ? await model.getVectorTilesSpec(db, collection.name) : [{"id": collection.name,"dataType":'vector'}];
  
  if (contentType == "json") reply.send(templates.collectionTileSet(baseurl, collection, layers));

};

async function getCollectionTile(req, reply) {
  const { collectionId, tileMatrix, tileRow, tileCol } = req.params;
  const { limit = 10000, properties = null } = req.query;
  const db = req.db || req.server.db;

  const {data_type} = modelCommon.getCollection(db, collectionId)

  let tile;
  if (data_type == 'features') tile = await helpers.getAsVectorTile(db, collectionId, tileMatrix, tileRow, tileCol, limit, properties)

  else tile = await model.getCollectionTile(db,collectionId, tileMatrix, tileRow, tileCol)
  if (tile) {
    if (typeof tile == "object") tile = Buffer.from(tile)
    const gzip = await helpers.isGzip(tile)

    reply
      .header('Content-Type', 'application/vnd.mapbox-vector-tile')
      .header('Content-Encoding', gzip ? 'gzip' : 'none')
      .send(tile)
  }
  else {
    reply.status(404).send()
  }

};



async function getCollectionMapTilesets(req, reply) {
  const { collectionId } = req.params;
  const { contentType } = req;
  const baseurl = req.params.dataset ? [req.server.baseurl, req.params.dataset].join("/") : req.server.baseurl;
  const db = req.db || req.server.db;

  if (contentType == "html") return reply.view("maptilesets", { baseurl, collectionId });

  let collection ={ 
    ...modelCommon.getCollection(db, collectionId),
    tileMatrixSetLimits: model.getTileMatrixSetLimits(db, collectionId)
  };

  reply.send(templates.collectionMapTileSets(baseurl, collection));
};


async function getCollectionMapTileset(req, reply) {
  const { collectionId } = req.params;
  const baseurl = req.params.dataset ? [req.server.baseurl, req.params.dataset].join("/") : req.server.baseurl;
  const db = req.db || req.server.db;

  let collection ={ 
    ...modelCommon.getCollection(db, collectionId),
    tileMatrixSetLimits: model.getTileMatrixSetLimits(db, collectionId)
  };

  reply.send(templates.collectionMapTileSet(baseurl, collection));
};

async function getCollectionMapTile(req, reply) {
  const { collectionId, tileMatrix, tileRow, tileCol } = req.params;
  const db = req.db || req.server.db;

  let tile = await model.getCollectionTile(db, collectionId, tileMatrix, tileRow, tileCol)
  if (tile) {
    if (typeof tile == "object") tile = Buffer.from(tile)
    const format = await filetypemime(tile);
    const gzip = await helpers.isGzip(tile)
    reply
      .header('Content-Type', format[0] || 'application/vnd.mapbox-vector-tile')
      .header('Content-Encoding', gzip ? 'gzip' : 'none')
      .send(tile)
  }
  else {
    reply.status(404).send()
  }

};



async function getTileMatrixSets(req, reply) {
  const { f } = req.query;
  const baseurl = req.params.dataset ? [req.server.baseurl, req.params.dataset].join("/") : req.server.baseurl;

  const contentType = f || req.accepts().type(['json', 'html']) || "json";
  if (contentType == "html") reply.view("tileMatrixSets", { baseurl });

  const tileMatrixSets = await model.getTileMatrixSets()
  reply.send(templates.tileMatrixSets(baseurl, tileMatrixSets))

};

async function getTileMatrixSet(req, reply) {
  const { f } = req.query;
  const { tileMatrixSetId } = req.params;
  const baseurl = req.params.dataset ? [req.server.baseurl, req.params.dataset].join("/") : req.server.baseurl;

  const tileMatrixSet = await model.getTileMatrixSet(tileMatrixSetId)
  const tileMatrices = await model.getTileMatrices(tileMatrixSetId);

  const templated = templates.tileMatrixSet(baseurl, tileMatrixSet, tileMatrices)

  const contentType = f || req.accepts().type(['json', 'html']) || "json";
  if (contentType == "html") reply.view("tileMatrixSet", { baseurl: baseurl, tileMatrixSetId, templated });

  reply.send(templated);

};


export {
  getTilesets,
  getTileset,
  getTile,

  getCollectionTilesets,
  getCollectionTileset,
  getCollectionTile,

  getCollectionMapTilesets,
  getCollectionMapTileset,
  getCollectionMapTile,

  getTileMatrixSets,
  getTileMatrixSet
}